refactor: split monolithic lib.rs into 11 focused modules with tests
Split the 2,489-line lib.rs monolith into 11 modules, each ≤600 lines
including inline tests. Added 86 native Rust tests covering all codec
functions (was: 5 tests).

Module layout:
  bitio.rs (495)    - BitWriter, BitReader, zigzag, extract_packed
  gorilla.rs (579)  - XOR Gorilla encode/decode/stats
  timestamp.rs (226)- DoD timestamp codec
  alp.rs (536)      - ALP core: exponent, encode, FoR decode
  alp_exc.rs (145)  - ALP exception decode (FoR + delta-FoR)
  delta_alp.rs (528)- Delta-ALP + WASM dispatch layer
  batch.rs (284)    - Batch XOR/ALP encode/decode
  range_decode.rs (227) - Fused ALP range decode
  alloc.rs (103)    - Scratch bump allocator
  interner.rs (241) - String interner (FNV)
  simd.rs (184)     - msToNs, quantizeBatch
  lib.rs (67)       - Thin shell: panic handler + mod declarations

Bug fixes discovered during testing:
  - extract_packed safe_limit: was `n - 8` (panics for small chunks);
    now computed from actual buffer length via packed_safe_limit()
  - ALP FoR range: `(max_int - min_int) as u64` overflowed for negative
    values; fixed with i128 intermediate cast
  - BitWriter::write_bits: count=0 caused shift overflow; added early return

Test infrastructure:
  - Shared static mut state (ALP_INTS, SCRATCH, etc.) protected by
    crate::test_lock::LOCK mutex for parallel test safety
  - All 86 tests pass in parallel (`cargo test`) and serial modes
